refactor: simplify useRetrieve by no using it in useContacts
