+++ src/shared/services/keychain.ts
@@ -64,7 +64,7 @@ export const retrieveStoredItem = async (key: string) => {
     storage: STORAGE_TYPE.AES,
   });
   if (!credentials) {
-    throw new Error(`No ${key} found`);
+    return undefined;
   }
   return credentials.password;
 };
